Extract to a common components folder shared parts of document playground to be used also in evaluation playground
