--- apps/web/src/app/(private)/projects/[projectId]/versions/[commitUuid]/documents/[documentUuid]/_components/DocumentEditor/Editor/Playground/components/index.ts
@@ -1,5 +1,0 @@
-export { FloatingElement } from './FloatingElement'
-export { StatusIndicator } from './StatusIndicator'
-export { StreamMessage } from './StreamMessage'
-export { Timer } from './Timer'
-export { TokenUsage } from './TokenUsage'
