+++ apps/web/src/app/(public)/share/d/[publishedDocumentUuid]/_components/Messages/AllMessages/index.tsx
@@ -10,7 +10,7 @@ import {
 } from '@latitude-data/compiler'
 import { ExpandMessages } from '../ExpandMessages'
 import { useToolContentMap } from 'node_modules/@latitude-data/web-ui/src/lib/hooks/useToolContentMap'
-import { StreamMessage } from '$/app/(private)/projects/[projectId]/versions/[commitUuid]/documents/[documentUuid]/_components/DocumentEditor/Editor/Playground/components'
+import { StreamMessage } from '$/components/PlaygroundCommon/StreamMessage'
 
 export function AllMessages({
   messages,
