+++ apps/web/src/components/PlaygroundCommon/StatusIndicator/FloatingElement/index.tsx
