+++ apps/web/src/components/PlaygroundCommon/StatusIndicator/index.tsx
