+++ apps/web/src/components/PlaygroundCommon/StreamMessage/index.tsx
