+++ apps/web/src/hooks/playgrounds/constants.ts
@@ -1,0 +1,4 @@
+import { COLLAPSED_BOX_HEIGHT } from '@latitude-data/web-ui/molecules/CollapsibleBox'
+
+export const PLAYGROUND_COLLAPSED_SIZE = COLLAPSED_BOX_HEIGHT * 2 + 12
+export const PLAYGROUND_GAP_PADDING = 26
