+++ apps/web/src/hooks/playgrounds/useStreamHandler.ts
